Cleanup of code to reduce duplication.

+++ src/main/java/nld/ede/runconnect/backend/dao/UserDAO.java
@@ -52,60 +52,25 @@ public class UserDAO implements IUserDAO
     /**
      * Follow or unfollow a user based on follower and followee ID.
      * This method only contains the business logic.
-     * @param follow Boolean value, if true follow, if false unfollow.
+     * @param wantsToFollow Boolean value, if true follow, if false unfollow.
      * @param followerId The ID of the person trying to follow a user.
      * @param followeeId The ID of the user to follow.
      * @return If the method was executed successful.
      */
-    public boolean toggleFollow(boolean follow, int followerId, int followeeId) throws SQLException
+    public boolean toggleFollow(boolean wantsToFollow, int followerId, int followeeId) throws SQLException
     {
         // If user is already following or already not following user, return out of method.
-        if (isFollowing(followerId, followeeId) == follow) {
+        if (isFollowing(followerId, followeeId) == wantsToFollow) {
             return false;
         }
 
-        return follow ? this.follow(followerId, followeeId) : this.unfollow(followerId, followeeId);
-    }
+        String query;
 
-    /**
-     * Follows a user based on follower and followee ID.
-     * @param followerId The ID of the follower.
-     * @param followeeId The ID of the followee.
-     * @return If the statement was successful.
-     * @throws SQLException Exception if SQL fails.
-     */
-    public boolean follow(int followerId, int followeeId) throws SQLException
-    {
-        String query = "INSERT INTO FOLLOWS (FOLLOWERID, FOLLOWEEID) VALUES (?, ?)";
-
-        try (Connection connection = dataSource.getConnection()) {
-            statement = connection.prepareStatement(query);
-            statement.setInt(1, followerId);
-            statement.setInt(2, followeeId);
-            resultSet = statement.executeQuery();
-
-            if (resultSet.next()) {
-                return true;
-            }
-        } catch (SQLException exception) {
-            throw exception;
-        } finally {
-            close(statement, resultSet);
+        if (wantsToFollow) {
+            query = "INSERT INTO FOLLOWS (FOLLOWERID, FOLLOWEEID) VALUES (?, ?)";
+        } else {
+            query = "DELETE FROM FOLLOWS WHERE FOLLOWERID = ? AND FOLLOWEEID = ?";
         }
-
-        return false;
-    }
-
-    /**
-     * Unfollows a user based on follower and followee ID.
-     * @param followerId The ID of the follower.
-     * @param followeeId The ID of the followee.
-     * @return If the statement was successful.
-     * @throws SQLException Exception if SQL fails.
-     */
-    public boolean unfollow(int followerId, int followeeId) throws SQLException
-    {
-        String query = "DELETE FROM FOLLOWS WHERE FOLLOWERID = ? AND FOLLOWEEID = ?";
 
         try (Connection connection = dataSource.getConnection()) {
             statement = connection.prepareStatement(query);
